refactor: collapse worktree barrel into worktree/index.ts
